fix(certora): fix failing rule that cancelled requests stay cancelled
The prover found a counter example where it starts with a request that
has an id `0`, which always results in a `SlotState.New`, resulting in
that request becoming `RequestState.Started` while it should be
`RequestState.Cancelled`.

We're fixing this by requiring that the `requestId != 0` and the
slot's client is `!= 0` address as well.

This will cause a reverting case in the prover, preventing it to find a
false positive.